extracted into two classes
