--- src/main/java/uk/ac/ebi/ampt2d/metadata/config/SecurityConfig.java
@@ -1,64 +1,0 @@
-/*
- *
- * Copyright 2018 EMBL - European Bioinformatics Institute
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- */
-package uk.ac.ebi.ampt2d.metadata.config;
-
-import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.oauth2.config.annotation.web.configuration.EnableResourceServer;
-import org.springframework.security.oauth2.config.annotation.web.configuration.ResourceServerConfigurerAdapter;
-
-public class SecurityConfig {
-
-    @ConditionalOnProperty(value = "security.enabled", havingValue = "true")
-    @Configuration
-    @EnableResourceServer
-    static class EnableSecurity extends ResourceServerConfigurerAdapter {
-
-        private static final String[] AUTH_WHITELIST = {
-                // -- swagger ui
-                "/v2/api-docs",
-                "/swagger-resources",
-                "/swagger-resources/**",
-                "/configuration/ui",
-                "/configuration/security",
-                "/swagger-ui.html",
-                "/webjars/**",
-                "/"
-        };
-
-        @Override
-        public void configure(HttpSecurity http) throws Exception {
-            http.authorizeRequests()
-                    .antMatchers(AUTH_WHITELIST).permitAll()
-                    .anyRequest().authenticated();
-
-        }
-    }
-
-    @ConditionalOnProperty(value = "security.enabled", havingValue = "false")
-    @Configuration
-    @EnableResourceServer
-    static class DisableSecurity extends ResourceServerConfigurerAdapter {
-
-        @Override
-        public void configure(HttpSecurity http) throws Exception {
-            http.authorizeRequests().anyRequest().permitAll();
-        }
-    }
-}
